+++ init_stack_b.c
@@ -6,7 +6,7 @@
 /*   By: yuocak <yuocak@student.42kocaeli.com.tr>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/20 17:50:25 by yuocak            #+#    #+#             */
-/*   Updated: 2025/03/24 00:46:58 by yuocak           ###   ########.fr       */
+/*   Updated: 2025/03/26 17:25:49 by yuocak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,13 +16,13 @@ void	set_target_stack_b(t_list *stack_a, t_list *stack_b, int a, int b)
 {
 	t_list	*current_a;
 	t_list	*target_node;
-	int		best_match_index;
+	long	best_match_index;
 
 	current_a = stack_a;
 	while (b-- != 0)
 	{
 		a = lst_len(stack_a);
-		best_match_index = INT_MAX;
+		best_match_index = LONG_MAX;
 		while (a-- != 0)
 		{
 			if (current_a->value > stack_b->value
@@ -33,7 +33,7 @@ void	set_target_stack_b(t_list *stack_a, t_list *stack_b, int a, int b)
 			}
 			current_a = current_a->next;
 		}
-		if (best_match_index == INT_MAX)
+		if (best_match_index == LONG_MAX)
 			stack_b->target_node = find_min(stack_a);
 		else
 			stack_b->target_node = target_node;
